split reducer by folder
