+++ project/src/store/reducer/root-reducer.ts
@@ -1,7 +1,7 @@
 import {combineReducers} from '@reduxjs/toolkit';
 import { EReducerNameSpace } from '../../types/const';
-import { hotelsData } from './hotels-data';
-import { userProcess } from './user-process';
+import { hotelsData } from './hotel-data/hotels-data';
+import { userProcess } from './user-process/user-process';
 
 export const rootReducer = combineReducers({
   [EReducerNameSpace.DATA]: hotelsData.reducer,
