+++ project/src/utils/mock.ts
@@ -2,7 +2,7 @@ import { system, address, lorem, helpers, name } from 'faker';
 import { HotelsData } from '../types/state';
 import { IHotel } from '../types/interfaces/hotel.interface';
 import { gerRandomNumber } from '../helpers/hepler';
-import { initialState } from '../store/reducer/hotels-data';
+import { initialState } from '../store/reducer/hotel-data/hotels-data';
 
 
 export const makeFakeInitialState: HotelsData = initialState;
